FIX: move initializer to separate file
